DOMA-1530 move chess editing from create to update form. Start update base input styles
