+++ apps/condo/domains/common/constants/style.js
@@ -112,10 +112,11 @@ const antGlobalVariables = {
     '@border-radius-base': '4px',
     '@white': white,
     '@black': black,
+    '@input-height-base': '32px',
     '@form-item-label-height': '@input-height-base',
     '@label-color': grey[2],
     '@input-border-color': inputBorderGrey,
-    '@input-bg': ultraLightGrey,
+    '@input-bg': whiteTranslucent,
     '@input-hover-border-color': sberGreen,
     '@select-background': ultraLightGrey,
     '@select-clear-background': ultraLightGrey,
