+++ apps/condo/domains/property/components/PropertyForm/index.tsx
@@ -1,6 +1,4 @@
 import React from 'react'
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
 import { UpdatePropertyForm } from './UpdatePropertyForm'
 import { CreatePropertyForm } from './CreatePropertyForm'
